[chore] #211 코드리뷰 반영
